Appease compile_et for objdir builds
The argument we pass to -p needs to be in the source tree, not
the object tree -- compile_et will not find the input files it
wants in the objdir tree.

For tbudb we can do this as is done on master, by just including
it in the local variable BUDB, but for tptserver and tvlserver
that is a rather invasive change.

This fixes the build with a separate objdir, which was broken by
commit 7c451fab11b65dd5e2860715b939e5de33c7645e.

Change-Id: If90f170ef3777a9987a0ba91747041abb89b2b65
Reviewed-on: http://gerrit.openafs.org/11435
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: D Brashear <shadow@your-file-system.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>